Drop column 'token' in table 17 (Identification Type)
The field IT - aka Identification Type - is defined as string and also
the example OCMF data shows this field with a string value.

So the token column in table 17 has no value at all, so let's drop it
from the specification.

While at, fix the wrong table numbering and adjust the wording
a little bit.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>